TDD (리팩토링)
